% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{associate}
\alias{associate}
\title{associate}
\usage{
associate(data, associationType)
}
\arguments{
\item{data}{data.frame containing EIA 'Plant.Code', 'Generator.ID' and 'Boiler.ID'}

\item{associationType}{character string 'bogen' indicating boiler-generator associations 
or 'bogencoo' indicating boiler-generator-cooling associations.}
}
\value{
data.frame with 'Plant.Code', 'Generator.ID', 'Boiler.ID' and 'Bogen'.  
'Bogen' is a concatenated string in the form of 'Plant.Code^boiler-generator-unit-number'
}
\description{
Generates concatenated boiler-generator (bogen) and 
boiler-generator-cooling (bogencoo) code applying the 
`igraph::graph_from_edgelist()` methods.  Bogen string used to calculate 
condenser duty at the boiler-generator level using the `condenserDuty` pacakge.
}
\examples{
#set arguments
dest<-tempdir()
eia_year<-2015
path_InputData_Metafile<-paste0(dest,.Platform$file.sep,"UserControlCrosswalk2015.xlsx")
path_EIAInputData<-paste0(dest,.Platform$file.sep,eia_year)
outputCSV<-FALSE
path_outputCSV<-dest

#use crosswalk2015 object
useStandardCrosswalk<-TRUE
prepCondenserDuty<-TRUE

#pull data from web
importThermoEIA::eia_webpull(eia_year,dest)


#run import using crosswalk2015
inputData.list<-importThermoEIA::import_EIAData(path_InputData_Metafile,path_EIAInputData,outputCSV,
                              path_outputCSV,prepCondenserDuty = TRUE,eia_year=eia_year,
                              useStandardCrosswalk=useStandardCrosswalk)
                              
associate(inputData.list$bogen,associationType="bogen")
}
