% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossCoolingID_cooling.R
\name{crossCoolingID_cooling}
\alias{crossCoolingID_cooling}
\title{crossCoolingID_cooling}
\usage{
crossCoolingID_cooling(
  bocoo,
  bogencoo.key,
  combogencoo_cooly_type,
  combogencoo_cooly_type_nukes,
  sheet4_key,
  nuke_combogencoo_key,
  boilerFuelData,
  gen_fuel_data
)
}
\description{
crossCoolingID_cooling
}
