% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIAplantAssociation-package.R
\docType{package}
\name{EIAplantAssociation-package}
\alias{EIAplantAssociation-package}
\title{Generate boiler-generator and boiler-generator-cooling associations}
\description{
\code{EIAplantAssociation} includes methods to generate boiler-generator and 
boiler-generator-cooling associations using data reported to the EIA on thermoelectric 
plants data found here ((https://www.eia.gov/electricity/).  Data used by the 
EIAplantAssociation package comes from the following EIA forms: Form-923, Form-860, 
Form-767, Form-906920, and Form-759. Boiler-generator  and boiler-generator-cooling 
associations are critical to calculating condenser duty at thermoelectric plants since
the data required to calculate fuel heat is reported by boiler and net generation is 
reported by generator, in order to calculate condenser duty The output of the 
EIAplantAssociation packages is formatted to be used by the condenserDuty package which 
calculates condenserDuty.  The EIAplantAssociation package expects the input data used 
for associations to be formatted according to the importThermoEIA package.
}
\details{
\tabular{ll}{
Package: \tab EIAplantAssociation \cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
https://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
Disclaimer: \tab This software is preliminary or provisional and is subject to revision. It is being 
provided to meet the need for timely best science. The software has not received final approval by the 
U.S. Geological Survey (USGS). No warranty, expressed or implied, is made by the USGS or the U.S. 
Government as to the functionality of the software and related material nor shall the fact of release 
constitute any such warranty. The software is provided on the condition that neither the USGS nor the 
U.S. Government shall be held liable for any damages resulting from the authorized or unauthorized use 
of the software.\cr
}

\tabular{ll}{

\code{\link{associate}} \tab Generates concatenated boiler-generator (bogen) code applying 
the `igraph::graph_from_edgelist()` methods.  Bogen string used to calculate 
condenser duty at the boiler-generator level using the `condenserDuty` pacakge.\cr \cr
}
}
\keyword{Thermoelectric}
