% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMissingPlants.R
\name{fileMissingPlants}
\alias{fileMissingPlants}
\title{fileMissingPlants}
\usage{
fileMissingPlants(plantList, inputData.list)
}
\arguments{
\item{plantList}{numeric vector of Plant.Codes to consider}

\item{inputData.list}{list object generated by the importThermoEIA::import_EIAData() function 
This data is pulled from the EIA website (https://www.eia.gov/electricity/) for the year 
of 2015 and has been run through the necessary formatting to be used by the EIAplantAssociation 
package.  inputData.list should be in the same format as the example `inputData.list2015` data
object.}
}
\value{
data.frame with the following columns \cr \cr
\itemize{\item    "Plant.Code"
          \item  "missingPlant.assoc" TRUE/FALSE indicating Plant.Code missing from `inputData.list$bogen`
          \item  "missingPlant.generator.data.860" TRUE/FALSE indicating Plant.Code missing from `inputData.list$gen_860` and `inputData.list$retiredGenerators`
          \item  "missingPlant.generation.data.923" TRUE/FALSE indicating Plant.Code missing from `inputData.list$generation.data`
          \item  "missingPlant.boilerFuelData.923" TRUE/FALSE indicating Plant.Code missing from `inputData.list$boilerFuelData`
          \item  "missingPlant.boilerDesignData.860" TRUE/FALSE indicating Plant.Code missing from `inputData.list$boilerDesignData`
          } \cr \cr
}
\description{
Finds Plant.Codes that are in the `plantList` but missing from one of the 
following tables \cr \cr
\itemize{  \item `inputData.list$bogen`
          \item `inputData.list$gen_860`
          \item `inputData.list$generation.data`
          \item `inputData.list$boilerFuelData`
          \item `inputData.list$boilerDesignData`
          \item `inputData.list$retiredGenerators`
          } \cr \cr
}
\examples{
#set arguments
dest<-tempdir()
eia_year<-2015
path_InputData_Metafile<-paste0(dest,.Platform$file.sep,"UserControlCrosswalk2015.xlsx")
path_EIAInputData<-paste0(dest,.Platform$file.sep,eia_year)
outputCSV<-FALSE
path_outputCSV<-dest

#use crosswalk2015 object
useStandardCrosswalk<-TRUE
prepCondenserDuty<-TRUE

#pull data from web
importThermoEIA::eia_webpull(eia_year,dest)


#run import using crosswalk2015
inputData.list<-importThermoEIA::import_EIAData(path_InputData_Metafile,path_EIAInputData,outputCSV,
                              path_outputCSV,prepCondenserDuty = TRUE,eia_year=eia_year,
                              useStandardCrosswalk=useStandardCrosswalk)

#get masterPlantList
data("masterPlantList")
plantList<-masterPlantList[masterPlantList$year==2015,]$Plant.Code

fileMissingPlants(plantList,inputData.list)
}
