% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatIDcol.R
\name{formatIDcol}
\alias{formatIDcol}
\title{formatIDcol}
\usage{
formatIDcol(df, idCol, flagColname, originalColname)
}
\arguments{
\item{df}{data.frame with "Plant.Code" (required name), ID (name set by `idCol`), and flag 
(name set by `flagCol`) columns.}

\item{idCol}{character string indicating which column is the ID column from which leading
zeros should be removed}

\item{flagColname}{character string indicating name of column in which to store the flag 
information indicating that the ID has been altered}

\item{originalColname}{character string indicating name of column in which to store the 
original un-altered EIA ID}
}
\value{
data.frame with IDs with special characters, spaces, and leading zeros removed and a 
       flag column indicating what type of removal has taken place ("removeSpace", 
       "removeChar" or "removeZero")
}
\description{
removes special characters, spaces, and leading zeros from character columns 
with IDs (i.e. Boiler.ID, Generator.ID, etc.).  This is necessary for boiler-generator 
associations because across EIA files there is a lack of consistency in ID names creating 
IDs that do not match with the these included. \cr \cr
}
\examples{
df<-data.frame(Plant.Code=c(9999,9999,9999,9999),
              ID=c("01","02","CT 02","GT-3"),
              flag=rep(NA,4))
formatIDcol(df,idCol="ID",flagColname = "flag",originalColname="orig_ID")
}
