% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualBogenReplace.R
\name{manualBogenReplace}
\alias{manualBogenReplace}
\title{manualBogenReplace}
\usage{
manualBogenReplace(
  outputData_path,
  fileName,
  data.list,
  replaceTables,
  lastReplace
)
}
\description{
manualBogenReplace
}
