\name{masterPlantList}
\alias{masterPlantList}
\docType{data}
\title{masterPlantList}
\description{
%%  ~~ A list of all plants recommended for plant association and/or condenserDuty analysis using the EIAplantAssociation and condenserDuty packages from 2008-2020. ~~
}
\usage{data("masterPlantList")}
\format{
  A data frame with 14903 observations on the following 2 variables.
  \describe{
    \item{\code{Plant.Code}}{a numeric vector indicating unique plant identifier from EIA}
    \item{\code{year}}{a numeric vector indicating the year in which a given Plant.Code should be including in EIAplantAssociation and/or condenserDuty packages}
  }
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(masterPlantList)
}
\keyword{datasets}
