% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.list.R
\name{named.list}
\alias{named.list}
\title{named.list}
\usage{
named.list(...)
}
\arguments{
\item{...}{unquoted object names to be compiled into a single list named with the object names.}
}
\value{
list object with all ... objects named with object names
}
\description{
Supports the creation of a list object by specifying unquoted object names in 
           the argument list. \\cr \\cr
}
\examples{
df<-data.frame(x=c(1,2,3),y=c(2,3,4))
myVector<-seq(1,200,1)
myList<-list(element1=data.frame(x=c(3,4,5)),setting="my custom list")
named.list(df,myVector,myList)
}
