% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLeadZero.R
\name{removeLeadZero}
\alias{removeLeadZero}
\title{removeLeadZero}
\usage{
removeLeadZero(df, idCol, flagCol)
}
\arguments{
\item{df}{data.frame with "Plant.Code" (required name), ID (name set by `idCol`), and flag 
(name set by `flagCol`) columns.}

\item{idCol}{character string indicating which column is the ID column from which leading
zeros should be removed}

\item{flagCol}{character string indicating which column name in which to store the flag 
information indicating that a zero was removed from the ID, can be output of 
the `formatIDcol()` function.}
}
\value{
data.frame with IDs with leading zeros removed and a flag column indicating
       when a removal has taken place by the string "removeZero"
}
\description{
removes leading zeros from character columns with IDs (i.e. Boiler.ID, 
Generator.ID, etc.).  This is necessary for boiler-generator associations because 
across EIA files there is a lack of consitency in ID names creating IDs that do not match
with the zeros included. \cr \cr
}
\examples{
df<-data.frame(Plant.Code=c(9999,9999,9999),ID=c("01","02","CT02"),flag=rep(NA,3))
removeLeadZero(df,idCol="ID",flagCol = "flag")
}
