% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumGenFuelData.R
\name{sumGenFuelData}
\alias{sumGenFuelData}
\title{sumGenFuelData}
\usage{
sumGenFuelData(gen_fuel_data, fuelHeatData, generation.data, analysisYear)
}
\arguments{
\item{gen_fuel_data}{data.frame stored in `inputData.list$gen_fuel_data` from EIA Form-923
"Page 1 Generation and Fuel Data" subset to only include plants that will be assigned a 
Plant.Code-Reported.Prime.Mover level bogen indicating that EIA Form-923 "Page 1 Generation and 
Fuel Data" should be used to calculate condenser duty.}

\item{fuelHeatData}{data.frame with Plant.Code, Reported.Prime.Mover_bf.923 
(Reported.Prime.Mover from the `boilerFuelData` table in `inputData.list`) and Fuel.Heat.(month)
from EIA Form-923 "Page 3 Boiler Fuel Data"}

\item{generation.data}{data.frame with Plant.Code, Reported.Prime.Mover 
(Reported.Prime.Mover from the `generation.data` table in `inputData.list`) and 
Net.Generation(month) from EIA Form-923 "Page 4 Generator Data"}
}
\value{
list of data.frames including:
#'\tabular{ll}{`gen_fuel_data` \tab data.frame of Fuel Heat and Net Generation at the 
Plant.Code-Reported.Prime.Mover level.  \cr \cr
`bf.923.ungrouped` \tab data.frame of Fuel Heat at the Plant.Code-boiler level for units not
selected in the `gen_fuel_data` argument and Fuel Heat at the Plant.Code-Reported.Prime.Mover
level for units selected in the `gen_fuel_data` argument.\cr \cr
`generation.data.line124` \tab data.frame of Net Generation at the Plant.Code-generator level
for units not selected in the `gen_fuel_data` argument and Net Generation at the 
Plant.Code-Reported.Prime.Mover level for units selected in the `gen_fuel_data` argument.\cr \cr
}
}
\description{
Calculates Fuel Heat and Net Generation at the Plant.Code-Reported.Prime.Mover
level using EIA Form-923 "Page 1 Generation and Fuel Data" and appends it to `fuelHeatData`
and `generation.data` \cr \cr
}
\examples{
#set arguments
dest<-tempdir()
eia_year<-2015
path_InputData_Metafile<-paste0(dest,.Platform$file.sep,"UserControlCrosswalk2015.xlsx")
path_EIAInputData<-paste0(dest,.Platform$file.sep,eia_year)
outputCSV<-FALSE
path_outputCSV<-dest

#use crosswalk2015 object
useStandardCrosswalk<-TRUE
prepCondenserDuty<-TRUE

#pull data from web
importThermoEIA::eia_webpull(eia_year,dest)


#run import using crosswalk2015
inputData.list<-importThermoEIA::import_EIAData(path_InputData_Metafile,path_EIAInputData,outputCSV,
                              path_outputCSV,prepCondenserDuty = TRUE,eia_year=eia_year,
                              useStandardCrosswalk=useStandardCrosswalk)

#extract elements of the inputData.list
list2env(inputData.list,envir = parent.frame())

#define months
months<-c("January","February","March","April","May",
          "June","July","August","September","October","November","December")
          
#calculate Fuel Heat at the Boiler.ID level
fuelHeatData<-boilerFuelData
for(m in months){
fuelHeatstr<-paste0("fuelHeatData<-fuelHeatData \%>\% dplyr::mutate(Fuel.Heat.",m,"=Quantity.Of.Fuel.Consumed.",m,"*MMbtu.Per.Unit.",m,")")
eval(parse(text=fuelHeatstr))
}

#add required column Reported.Prime.Mover_bf.923
fuelHeatData$Reported.Prime.Mover_bf.923<-fuelHeatData$Reported.Prime.Mover

#set analysis Year
analysisYear<-2015

sumGenFuelData(gen_fuel_data,fuelHeatData, generation.data,analysisYear)
}
