% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeColumnLegend_metafile.R
\name{writeColumnLegend_metafile}
\alias{writeColumnLegend_metafile}
\title{writeColumnLegend_metafile}
\usage{
writeColumnLegend_metafile(outputDir)
}
\arguments{
\item{outputDir}{character string indicating directory in which to save the 
"ColumnLegend_metafile.xlsx" formatted file.}
}
\description{
creates a formatted Microsoft Excel® file with the flags described in the 
`data(ColumnLegend_metafile)` object associated with the creation of boiler-generator 
(bogen) associations using the `bogen_associate()` function. \cr \cr
}
\examples{
outputDir<-tempDir()
writeColumnLegend_metafile(outputDir)
}
